add pp unittest
